No more virtual files
